Split http_server into server + base class
The base class contains the common functionality that will also be used
by the wallet rpc server.
